Moved ensure_cugraph_obj() to utils.
